move docker tasks from controller into DockerWrapper lib
