+++ lib/DockerWrapper.js
@@ -1,5 +1,6 @@
 const Docker = require('dockerode');
 const Machine = require('docker-machine');
+const Node = require('../server/models').Node;
 const fs = require('fs');
 const path = require('path');
 const uuidv1 = require('uuid/v1')
@@ -18,72 +19,53 @@ CMD ["bundle", "exec", "puma", "-t", "5:5", "-p", "4567"]`;
 };
 
 module.exports = {
-  buildDroplet(app) {
-    return new Promise((resolve, reject) => {
-      console.log('Building droplet...');
-
-      const options = {
-        'digitalocean-access-token': process.env.ACCESS_TOKEN,
-      };
-
-      Machine.create(app.dropletName, 'digitalocean', options, (err) => {
-        if (err) throw err;
-        resolve(app);
-      });
-    });
-  },
-
-  buildDockerfile() {
+  buildDockerfile: (file) => {
     return (app) => {
       return new Promise((resolve, reject) => {
         console.log('Building Dockerfile...')
         fs.writeFileSync(`./${app.path}/Dockerfile`, dockerfileContent(file));
+        console.log('Built Dockerfile!');
         resolve(app);
       });
     };
   },
 
-  buildAndRunContainer() {
+  buildImage: (app) => {
     return new Promise(async (resolve, reject) => {
-      console.log('Building + running container on droplet...');
-      const machine = new Machine(app.dropletName);
+      Node.findOne({ where: { manager: true }}).then((manager) => {
+        const machine = new Machine(manager.name);
 
-      machine.env({ parse: true }, (err, result) => {
-        const certPath = result.DOCKER_CERT_PATH;
-        const hostWithPort = result.DOCKER_HOST.split('//')[1];
-        const host = hostWithPort.split(':')[0];
-        const port = hostWithPort.split(':')[1];
+        machine.env({ parse: true }, (err, result) => {
+          if (err) { console.log(err); }
+          const certPath = result.DOCKER_CERT_PATH;
+          const hostWithPort = result.DOCKER_HOST.split('//')[1];
+          const host = hostWithPort.split(':')[0];
+          const port = hostWithPort.split(':')[1];
 
-        const options = {
-          socketPath: undefined,
-          host: host,
-          port: Number(port),
-          ca: fs.readFileSync(certPath + '/ca.pem'),
-          cert: fs.readFileSync(certPath + '/cert.pem'),
-          key: fs.readFileSync(certPath + '/key.pem'),
-        };
+          const options = {
+            socketPath: undefined,
+            host: host,
+            port: Number(port),
+            ca: fs.readFileSync(certPath + '/ca.pem'),
+            cert: fs.readFileSync(certPath + '/cert.pem'),
+            key: fs.readFileSync(certPath + '/key.pem'),
+          };
 
-        const docker = new Docker(options);
+          const docker = new Docker(options);
+          const tarStream = tar.pack(app.path);
 
-        var tarStream = tar.pack(app.path);
-        docker.buildImage(tarStream, {
-          t: `${app.title}:latest`,
-        }, function(error, output) {
-          if (error) {
-            return console.error(error);
-          }
-          output.pipe(process.stdout);
-          output.on('end', function () {
-            // send notification to client that build is ready
-            var container = docker.createContainer({
-              Image: app.title + ':latest',
-              PortBindings: { "4567/tcp": [{ HostPort: "80" }] }
-            }).then(container => {
-              return container.start();
-            }).then(container => {
+          console.log('Building image...');
+          docker.buildImage(tarStream, {
+            t: `${app.title}:latest`,
+          }, function(error, output) {
+            if (error) {
+              return console.error(error);
+            }
+            output.pipe(process.stdout);
+            output.on('end', function () {
+              // send notification to client that build is ready
+              console.log('Image built!');
               resolve(app);
-            }).catch(error => {
-              throw error;
             });
           });
         });
@@ -91,16 +73,99 @@ module.exports = {
     });
   },
 
-  saveIpAddress() {
+  createService: (app) => {
     return new Promise((resolve, reject) => {
-      const machine = new Machine(app.dropletName);
+      Node.findOne({ where: { manager: true }}).then((manager) => {
+        const machine = new Machine(manager.name);
 
-      machine.inspect((err, result) => {
-        if (err) throw err;
+        machine.env({ parse: true }, (err, result) => {
+          if (err) { console.log(err); }
 
-        const ipAddress = result.driver.ipAddress;
-        app.update({ ipAddress }).then((app) => resolve(app));
+          const certPath = result.DOCKER_CERT_PATH;
+          const hostWithPort = result.DOCKER_HOST.split('//')[1];
+          const host = hostWithPort.split(':')[0];
+          const port = hostWithPort.split(':')[1];
+
+          const options = {
+            socketPath: undefined,
+            host: host,
+            port: Number(port),
+            ca: fs.readFileSync(certPath + '/ca.pem'),
+            cert: fs.readFileSync(certPath + '/cert.pem'),
+            key: fs.readFileSync(certPath + '/key.pem'),
+          };
+
+          const docker = new Docker(options);
+
+          const serviceParams = {
+            "Name": "web",
+            "TaskTemplate": {
+              "ContainerSpec": {
+                "Image": `${app.title}:latest`,
+              },
+              "Networks": [
+                { "Target": `${app.title}_default` }
+              ]
+            },
+            "EndpointSpec": {
+              "Ports": [
+                {
+                  "Protocol": "tcp",
+                  "PublishedPort": 8080,
+                  "TargetPort": 4567
+                }
+              ]
+            },
+          };
+
+          console.log('Creating service...');
+          docker.createService(serviceParams, (error, result) => {
+            if (error) { console.log(error); }
+            console.log('Service created!');
+            resolve(app);
+          });
+        });
       });
     });
-  }
+  },
+
+  createNetwork: (app) => {
+    return new Promise(async (resolve, reject) => {
+      Node.findOne({ where: { manager: true }}).then((manager) => {
+        console.log('Creating network...');
+        const machine = new Machine(manager.name);
+
+        machine.env({ parse: true }, (err, result) => {
+          if (err) { console.log(err); }
+          const certPath = result.DOCKER_CERT_PATH;
+          const hostWithPort = result.DOCKER_HOST.split('//')[1];
+          const host = hostWithPort.split(':')[0];
+          const port = hostWithPort.split(':')[1];
+
+          const options = {
+            socketPath: undefined,
+            host: host,
+            port: Number(port),
+            ca: fs.readFileSync(certPath + '/ca.pem'),
+            cert: fs.readFileSync(certPath + '/cert.pem'),
+            key: fs.readFileSync(certPath + '/key.pem'),
+          };
+
+          const docker = new Docker(options);
+
+          const networkOptions = {
+            "Name": app.title + '_default',
+            "Driver": "overlay"
+          };
+
+          console.log('Creating network...');
+          docker.createNetwork(networkOptions, (error, result) => {
+            if (error) { console.log(error); }
+             console.log('Network created!');
+            resolve(app);
+          });
+        });
+      });
+    });
+  },
 };
